* utils/predicates/InstanceOf.java: My first default predicate. More to come.
* utils/gui/SortedListModel.java: Having a sorted list model is a
really nifty idea, so I wrote one.  It's being used in JavaDome
for the nodes in the main editor panel.

* utils/event/CollectionEvent.java: Added collection events for
collections that you can listen to.

* utils/algorithms/Finding.java: Added to help replace JGL.

* utils/algorithms/Filtering.java: Java's collections framework is
just too far from JGL.  Need to enhance it with Filters.

* utils/UnaryPredicate.java: Needed for Finding and Filtering classes.

* utils/Unique.java: Added default comparator for use when
creating total order comparators.

* utils/DefaultUnique.java: Added a correct equals method.

* utils/CollectionUtils.java: Needed instanceof check for objects
in a container, particularly for assertions.

* utils/Bean.java: Now follows standard naming conection where
interface is the name used and Default is prepended for default class.

* utils/gui/BetterBoxLayout.java: Added for better layout.  Swing's box
doesn't size so good.